requirements.txt
asf-search>=7.0.0
hyp3-sdk>=4.0.0
rasterio>=1.3.0
geopandas>=0.14.0
matplotlib>=3.7.0
folium>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
shapely>=2.0.0
geojson>=3.1.0
click>=8.1.7
tqdm>=4.66.0
